Remove unnecessary type union varargs
Use of type-union varargs allows creation of new instances of Combatants
to be slightly prettier at the expense of requiring extra traversals and
type inspection in order to separate them out, which isn't extensible.
Often, we clearly already have distinct sets of tools/traits which have
to be merged, only to be unmerged later.