Add annotation to verify that a method actually overrides something
BatteryStatus had a bug where it was trying to override a method, but it
used the wrong name, and the compiler couldn't catch that it wasn't
overriden. This uses the experimental annotations mechanism to add an
annotation similar to @Override in Java, and adds these annotations to
all overriden methods from AbstractStatus.