Implement LocalStorage for settings, switch volume to use Settings
Adds a Settings class that Save can use to cache settings. Currently,
the volume slider is the only available setting. Also adds
SettingsOptions, which is the type of the JSON object that's parsed from
and dumped to local storage.

Switches volume controls to use the Settings volume controls, instead of
setting directly to SoundManager.

The API here could use a little work; the SettingsOptions should really
be able to be derived from Settings automatically, and it's confusing
where load/save calls should occur, as well as whether one should call
Settings.setVolumePercent or SoundManager.setVolume, the latter of which
should only ever be called inside of the former.

Addresses #31.